test(components): added test in order form
