--- src/modules/recordSaga.js
@@ -1,15 +1,0 @@
-import { runSaga } from 'redux-saga';
-
-export async function recordSaga(saga, initialAction) {
-    const dispatched = [];
-
-    await runSaga(
-        {
-            dispatch: (action) => dispatched.push(action),
-        },
-        saga,
-        initialAction
-    ).done;
-
-    return dispatched;
-}
